Elide casts for stores to known globals

+++ tests/end_to_end/complicated_context/irix-g-out.c
@@ -5,8 +5,8 @@ s16 test(struct SomeStruct *arg, u8 should, f64 union_arg, ? union_arg_unk4, ...
     temp_a1 = should & 0xFF;
     sp0 = union_arg;
     if (temp_a1 != 0) {
-        globalf = (f32) arg->float_field;
-        globali = (s32) arg->int_field;
+        globalf = arg->float_field;
+        globali = arg->int_field;
         arg->data_field.char_innerfield = temp_a1;
     } else {
         arg->pointer_field = NULL;
